[ZhangJin] feat: add implement and test for add a trainee

+++ src/main/java/com/example/demo/repository/TrainerRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface TrainerRepository extends JpaRepository<TrainerEntity,Long> {
+public interface TrainerRepository extends JpaRepository<TrainerEntity,Integer> {
 
 }
